--- main.css
@@ -1,5 +1,0 @@
-.card {
-    padding: 1rem;
-    border-radius: 1rem;
-    overflow: hidden;
-}
