+++ pages/Prediction/Prediction.md
@@ -1,0 +1,1 @@
+# Prediction
