+++ pages/Prediction/Prediction.py
@@ -1,0 +1,11 @@
+"""
+A page of the application.
+Page content is imported from the Prediction.md file.
+
+Please refer to https://docs.taipy.io/en/latest/manuals/gui/pages for more details.
+"""
+
+import taipy as tp
+from taipy.gui import Markdown
+
+Prediction = Markdown("pages/Prediction/Prediction.md")
