+++ pages/__init__.py
@@ -1,0 +1,3 @@
+from .root import root_page
+from .Prediction.Prediction import Prediction
+from .Drift.Drift import Drift
